Update the current tests for Designer
